apps/api/requirements.txt
fastapi==0.135.3
python-multipart==0.0.26
uvicorn[standard]==0.30.6
PyYAML==6.0.3
httpx==0.28.1
ansible==13.3.0
ruamel.yaml==0.19.1
bcrypt==5.0.0
paramiko==4.0.0
pykeepass==4.1.1.post1
pycryptodomex==3.23.0
psycopg[binary]==3.3.3
